+++ src/js-framework/lib/vm/instance/scope.js
@@ -116,6 +116,7 @@ exports._proxy = function (key) {
  */
 
 exports._unproxy = function (key) {
+  /* istanbul ignore next */
   delete this[key]
 }